requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
flask-cors==4.0.0
python-magic==0.4.27
ffmpeg-python==0.2.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
openai-whisper==20250625
torch==2.8.0
more-itertools==10.8.0
googletrans==4.0.0rc1
deep-translator==1.11.4
requests==2.32.5
yt-dlp
celery==5.5.3
redis==6.4.0
reportlab==4.4.3